requirements.txt
psycopg2-binary==2.9.9
Faker==24.4.0
pandas==2.2.1
numpy==1.26.4
python-dotenv==1.0.1
sqlparse==0.4.4
sqlalchemy==2.0.28
pytest==8.1.1
snowflake-connector-python==3.7.1
schedule==1.2.2
pymongo==4.6.3
